Actually compresses dmis when saving them
We've been using the write_to wrapper function, which doesn't allow us
to pass in a compression level.
This means we've been using the default, which is for some reason the
fastest/worst for file size.

This commit instead "inlines" the write_to function, and uses the Best
compression method (slowest/lowest file size)

This strategy reduces a large file I tested (screen_cyborg.dmi) from
82kb to 21kb (byond's output if you're curious is about 26kb)

I've also added saving to the test suite, to ensure the path gets walked
and to allow for visual inspection if changes are made